services/ops/scripts/orchestration/requirements-airflow.txt
loguru==0.7.2
mlflow==2.17.2
pandas
scikit-learn
xgboost
boto3
clickhouse-connect==0.8.18
clickhouse-driver
apache-airflow-providers-apache-spark
apache-airflow-providers-openlineage>=1.8.0
apache-airflow-providers-docker
# sqlalchemy<2.0.0
# dbt-clickhouse==1.5.0
# dbt-core==1.5.11
# dbt-extractor==0.4.1
dbt-core
dbt-clickhouse